+++ session.go
@@ -10,13 +10,13 @@ type (
 
 	// Session information about the current transmission session
 	Session struct {
-		AltSpeedDown    int  `json:"alt-speed-down"`
-		AltSpeedEnabled bool `json:"alt-speed-enabled"`
+		AltSpeedDown    float64 `json:"alt-speed-down"`
+		AltSpeedEnabled bool    `json:"alt-speed-enabled"`
 		//Alt_speed_time_begin   int    `json:"alt-speed-time-begin"`
 		//Alt_speed_time_day     int    `json:"alt-speed-time-day"`
 		//Alt_speed_time_enabled bool   `json:"alt-speed-time-enabled"`
 		//Alt_speed_time_end     int    `json:"alt-speed-time-end"`
-		AltSpeedUp int `json:"alt-speed-up"`
+		AltSpeedUp float64 `json:"alt-speed-up"`
 		//Blocklist_enabled       bool   `json:"blocklist-enabled"`
 		//Blocklist_size          int    `json:"blocklist-size"`
 		//Blocklist_url           string `json:"blocklist-url"`
@@ -50,9 +50,9 @@ type (
 		SeedQueueSize         int     `json:"seed-queue-size"`
 		SeedRatioLimit        float64 `json:"seedRatioLimit"`
 		SeedRatioLimited      bool    `json:"seedRatioLimited"`
-		SpeedLimitDown        int     `json:"speed-limit-down"`
+		SpeedLimitDown        float64 `json:"speed-limit-down"`
 		SpeedLimitDownEnabled bool    `json:"speed-limit-down-enabled"`
-		SpeedLimitUp          int     `json:"speed-limit-up"`
+		SpeedLimitUp          float64 `json:"speed-limit-up"`
 		SpeedLimitUpEnabled   bool    `json:"speed-limit-up-enabled"`
 		//Start_added_torrents         bool   `json:"start-added-torrents"`
 		//Trash_original_torrent_files bool   `json:"trash-original-torrent-files"`
